5주차 1차 배포

+++ client/src/container/TransactionUpdateModal/index.tsx
@@ -2,7 +2,7 @@ import Modal from '@/components/common/Modal';
 import ModalHeader from '@/components/transaction/ModalHeader';
 import { RootState } from '@/modules';
 import { toggleModalOff } from '@/modules/updateModal';
-import React, { useCallback, useState, useReducer, useEffect } from 'react';
+import React, { useCallback, useState, useReducer, useEffect, useMemo } from 'react';
 import CategoryDTO from '@/commons/dto/category';
 import PaymentDTO from '@/commons/dto/payment';
 import { useDispatch, useSelector } from 'react-redux';
@@ -11,21 +11,25 @@ import { UpdateTransactionRequest } from '@/commons/types/transaction';
 import ModalInput from '@/components/transaction/ModalInput';
 import CustomSelectInput from '@/components/common/forms/CustomSelectInput';
 import CustomButton from '@/components/common/buttons/CustomButton';
+import checkValidation from '@/libs/checkValidation';
 import TransactionRequestDTO from '@/commons/dto/transaction-request';
 import { deleteTransactionThunk, updateTransactionThunk } from '@/modules/transaction';
 import * as S from './styles';
 
+const MODAL_LIST_ARR = ['tradeAt', 'description', 'amount', 'pid', 'cid', 'isIncome'];
+
 const TransactionUpdateModal = (): JSX.Element => {
   const [isIncome, setIsIncome] = useState(false);
+  const [validation, setValidation] = useState(new Set(MODAL_LIST_ARR));
   const { payment, category } = useSelector((state: RootState) => state);
   const { toggle, data } = useSelector((state: RootState) => state.updateModal);
-  const categoryList = category.data.map((c) => new CategoryDTO(c));
-  const paymentList = payment.data.map((p) => new PaymentDTO(p));
-
+  const categoryList = useMemo(() => category.data.map((c) => new CategoryDTO(c)), [category]);
+  const paymentList = useMemo(() => payment.data.map((p) => new PaymentDTO(p)), [payment]);
   const dispatch = useDispatch();
   const toggleModal = useCallback(() => {
     dispatch(toggleModalOff());
-  }, [dispatch]);
+    setValidation(new Set(MODAL_LIST_ARR));
+  }, []);
 
   const onChangeReducer = (state: UpdateTransactionRequest, action: UpdateTransactionRequest) => {
     return action;
@@ -34,8 +38,20 @@ const TransactionUpdateModal = (): JSX.Element => {
   const [updatedTransaction, infoDispatch] = useReducer(onChangeReducer, {
     tid: data?.tid,
   } as UpdateTransactionRequest);
+
   const onChangeInput = (e: React.ChangeEvent<HTMLInputElement>) => {
-    infoDispatch({ ...updatedTransaction, [e.target.name]: e.target.value });
+    if (checkValidation(e.target.name, e.target.value)) {
+      infoDispatch({ ...updatedTransaction, [e.target.name]: e.target.value });
+      validation.add(e.target.name);
+      setValidation(new Set([...validation]));
+    } else {
+      validation.delete(e.target.name);
+      setValidation(new Set([...validation]));
+    }
+    if ((e.target.name === 'description' || e.target.name === 'amount') && e.target.value === '') {
+      validation.delete(e.target.name);
+      setValidation(new Set([...validation]));
+    }
   };
 
   useEffect(() => {
@@ -56,13 +72,17 @@ const TransactionUpdateModal = (): JSX.Element => {
     const newTransactionDTO = new TransactionRequestDTO(updatedTransaction);
     dispatch(updateTransactionThunk(newTransactionDTO));
     dispatch(toggleModalOff());
-  }, [dispatch, updatedTransaction, data]);
+  }, [updatedTransaction, data]);
 
   const deleteTransaction = useCallback(() => {
     if (!data) return;
-    dispatch(deleteTransactionThunk(data.tid));
-    dispatch(toggleModalOff());
-  }, [dispatch, data]);
+    if (window.confirm('삭제 하시겠습니까?')) {
+      dispatch(deleteTransactionThunk(data.tid));
+      dispatch(toggleModalOff());
+    } else {
+      toggleModalOff();
+    }
+  }, [data]);
 
   return (
     <>
@@ -117,7 +137,11 @@ const TransactionUpdateModal = (): JSX.Element => {
             <CustomButton color="white" onClickEvent={deleteTransaction}>
               삭제
             </CustomButton>
-            <CustomButton color="blue" onClickEvent={updateTransaction}>
+            <CustomButton
+              isValid={validation.size > 5}
+              color="blue"
+              onClickEvent={updateTransaction}
+            >
               저장
             </CustomButton>
           </S.ModalFooter>
